`protect()` overload for CF types is too permissive https://bugs.webkit.org/show_bug.cgi?id=306615
Reviewed by Geoffrey Garen.

`protect()` overload for CF types is too permissive and can get called
for non-CFTypes. This is because `CFTypeRef` is a typedef to `void*`.

To address the issue, I am introduced a new `IsCFType` (Similar to the
existing `IsNSType`) which matches CFTypes based on whether or not
CFTypeTrait exists for the type.

* Source/WTF/WTF.xcodeproj/project.pbxproj:
* Source/WTF/wtf/PlatformMac.cmake:
* Source/WTF/wtf/RetainPtr.h:
* Source/WTF/wtf/cf/CFTypeTraits.h: Copied from Source/WTF/wtf/cf/TypeCastsCF.h.
* Source/WTF/wtf/cf/TypeCastsCF.h:

Canonical link: https://commits.webkit.org/306493@main
