+++ Source/WTF/wtf/cf/TypeCastsCF.h
@@ -30,14 +30,13 @@
 #include <objc/objc.h>
 #include <wtf/Assertions.h>
 #include <wtf/RetainPtr.h>
+#include <wtf/cf/CFTypeTraits.h>
 
 #ifndef CF_BRIDGED_TYPE
 #define CF_BRIDGED_TYPE(T)
 #endif
 
 namespace WTF {
-
-template <typename> struct CFTypeTrait;
 
 // Use dynamic_cf_cast<> instead of checked_cf_cast<> when actively checking CF types,
 // similar to dynamic_cast<> in C++. Be sure to include a nullptr check.
@@ -100,33 +99,8 @@ inline id bridge_id_cast(CFTypeRef object)
 
 } // namespace WTF
 
-#define WTF_DECLARE_CF_TYPE_TRAIT(ClassName) \
-template <> \
-struct WTF::CFTypeTrait<ClassName##Ref> { \
-    static inline CFTypeID typeID(void) { return ClassName##GetTypeID(); } \
-};
-
-WTF_DECLARE_CF_TYPE_TRAIT(CFArray);
-WTF_DECLARE_CF_TYPE_TRAIT(CFBoolean);
-WTF_DECLARE_CF_TYPE_TRAIT(CFData);
-WTF_DECLARE_CF_TYPE_TRAIT(CFDictionary);
-WTF_DECLARE_CF_TYPE_TRAIT(CFNumber);
-WTF_DECLARE_CF_TYPE_TRAIT(CFString);
-WTF_DECLARE_CF_TYPE_TRAIT(CFURL);
+// CTFontDescriptor requires CoreText, so we declare its trait here rather than in CFTypeTraits.h
 WTF_DECLARE_CF_TYPE_TRAIT(CTFontDescriptor);
-
-#define WTF_DECLARE_CF_MUTABLE_TYPE_TRAIT(ClassName, MutableClassName) \
-template <> \
-struct WTF::CFTypeTrait<MutableClassName##Ref> { \
-    static inline CFTypeID typeID(void) { return ClassName##GetTypeID(); } \
-};
-
-WTF_DECLARE_CF_MUTABLE_TYPE_TRAIT(CFArray, CFMutableArray);
-WTF_DECLARE_CF_MUTABLE_TYPE_TRAIT(CFData, CFMutableData);
-WTF_DECLARE_CF_MUTABLE_TYPE_TRAIT(CFDictionary, CFMutableDictionary);
-WTF_DECLARE_CF_MUTABLE_TYPE_TRAIT(CFString, CFMutableString);
-
-#undef WTF_DECLARE_CF_MUTABLE_TYPE_TRAIT
 
 using WTF::bridgeCFCast;
 using WTF::bridge_id_cast;
